Fixed import signatures from restructuring.
