+++ Blankly/__init__.py
@@ -16,10 +16,16 @@
     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
 
+from Blankly.exchanges.interfaces.Coinbase_Pro.Coinbase_Pro import Coinbase_Pro as Coinbase_Pro
+from Blankly.exchanges.interfaces.Binance.Binance import Binance as Binance
 from Blankly.exchanges.interfaces.Alpaca.Alpaca import Alpaca as Alpaca
-from Blankly.strategy.blankly_bot import BlanklyBot
+from Blankly.exchanges.interfaces.Paper_Trade.Paper_Trade import PaperTrade
 
+from Blankly.exchanges.managers.ticker_manager import TickerManager as TickerManager
 from Blankly.exchanges.managers.orderbook_manager import OrderbookManger as OrderbookManager
+from Blankly.exchanges.managers.general_stream_manager import GeneralManager as GeneralManager
+from Blankly.exchanges.interfaces.abc_currency_interface import ICurrencyInterface as Interface
+from Blankly.strategy.blankly_bot import BlanklyBot
 import Blankly.utils.utils as utils
 from Blankly.utils.scheduler import Scheduler
 import Blankly.indicators as indicators
