+++ Blankly/exchanges/orders/order.py
@@ -15,7 +15,6 @@
     You should have received a copy of the GNU Lesser General Public License
     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
-
 from Blankly.utils.utils import pretty_print_JSON
