+++ tests/exchanges/test_interface_homogeneity.py
@@ -58,11 +58,11 @@ class InterfaceHomogeneity(unittest.TestCase):
         cls.Interfaces.append(cls.Binance_Interface)
 
         # Alpaca definition and appending
-        cls.Alpaca = Blankly.Alpaca(portfolio_name="alpaca test portfolio",
-                                    keys_path='./tests/config/keys.json',
-                                    settings_path="./tests/config/settings.json")
-        cls.Alpaca_Interface = cls.Alpaca.get_interface()
-        cls.Interfaces.append(cls.Alpaca_Interface)
+        # cls.Alpaca = Blankly.Alpaca(portfolio_name="alpaca test portfolio",
+        #                             keys_path='./tests/config/keys.json',
+        #                             settings_path="./tests/config/settings.json")
+        # cls.Alpaca_Interface = cls.Alpaca.get_interface()
+        # cls.Interfaces.append(cls.Alpaca_Interface)
 
         # Paper trade wraps binance
         cls.Paper_Trade = Blankly.PaperTrade(cls.Binance)
